port chapter 21

+++ Chapter 17 Cube Mapping/CubeMap/Sky.cpp
@@ -28,7 +28,7 @@ Sky::Sky(ID3D11Device* device, const std::wstring& cubemapFilename, float skySph
 
     D3D11_BUFFER_DESC vbd;
     vbd.Usage = D3D11_USAGE_IMMUTABLE;
-	vbd.ByteWidth = sizeof(XMFLOAT3) * vertices.size();
+	vbd.ByteWidth = (UINT)(sizeof(XMFLOAT3) * vertices.size());
     vbd.BindFlags = D3D11_BIND_VERTEX_BUFFER;
     vbd.CPUAccessFlags = 0;
     vbd.MiscFlags = 0;
@@ -40,7 +40,7 @@ Sky::Sky(ID3D11Device* device, const std::wstring& cubemapFilename, float skySph
     HR(device->CreateBuffer(&vbd, &vinitData, &mVB));
 	
 
-	mIndexCount = sphere.Indices.size();
+	mIndexCount = (UINT)(sphere.Indices.size());
 
 	D3D11_BUFFER_DESC ibd;
     ibd.Usage = D3D11_USAGE_IMMUTABLE;
@@ -76,13 +76,10 @@ void Sky::Draw(ID3D11DeviceContext* dc, const Camera& camera)
 	// center Sky about eye in world space
 	XMFLOAT3 eyePos = camera.GetPosition();
 	XMMATRIX T = XMMatrixTranslation(eyePos.x, eyePos.y, eyePos.z);
-
-
 	XMMATRIX WVP = XMMatrixMultiply(T, camera.ViewProj());
 
 	Effects::SkyFX->SetWorldViewProj(WVP);
 	Effects::SkyFX->SetCubeMap(mCubeMapSRV);
-
 
 	UINT stride = sizeof(XMFLOAT3);
     UINT offset = 0;
@@ -90,7 +87,7 @@ void Sky::Draw(ID3D11DeviceContext* dc, const Camera& camera)
 	dc->IASetIndexBuffer(mIB, DXGI_FORMAT_R16_UINT, 0);
 	dc->IASetInputLayout(InputLayouts::Pos);
 	dc->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-	
+
 	D3DX11_TECHNIQUE_DESC techDesc;
     Effects::SkyFX->SkyTech->GetDesc( &techDesc );
